requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.3.2
contourpy==1.3.0
cycler==0.12.1
Django==4.2.19
djangorestframework==3.15.2
djongo==1.3.6
dnspython==2.6.1
fonttools==4.53.1
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.4.5
kiwisolver==1.4.7
Markdown==3.7
matplotlib==3.9.2
numpy==2.0.2
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pydub==0.25.1
pymongo==4.9.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
six==1.16.0
soupsieve==2.6
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
zipp==3.20.2
